Move ScenarioExtra class to common library
