+++ include/boe/scenario_extra.hpp
