+++ src/Windows Code Release 3/Scen Ed/CMakeLists.txt
@@ -11,7 +11,6 @@ add_executable (Scen-Ed WIN32
 	townout_stub.cpp
 	global.cpp
 	buttonmg.cpp
-	../scenario_ext.cpp
 	scened_globals.cpp
 
 	BLSCENED.RC
