+++ src/Windows Code Release 3/dlgutils.cpp
@@ -25,7 +25,6 @@
 #include "graphutl.h"
 #include "boe/utility.hpp"
 #include "boe/item.hpp"
-#include "scenario_ext.hpp"
 #include "game_globals.hpp"
 
 #define	NUM_HINTS	30
