+++ src/Windows Code Release 3/game_globals.hpp
@@ -1,7 +1,7 @@
 #pragma once
 
 #include "boe/scenario.hpp"
-#include "scenario_ext.hpp"
+#include "boe/scenario_extra.hpp"
 
 extern const Scenario scenario;
 extern const ScenarioExtra scenario_ext;
