+++ src/Windows Code Release 3/text.cpp
@@ -18,7 +18,6 @@
 #include "graphutl_helpers.hpp"
 #include "boe/hacks.hpp"
 #include "boe/item.hpp"
-#include "scenario_ext.hpp"
 #include "game_globals.hpp"
 
 static const std::array m_mage_sp{"Spark","Minor Haste","Strength","Flame Cloud","Flame",
